requirements.txt
streamlit
ultralytics
torch
numpy
Pillow
opencv-python-headless
